# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MAWIBO Super Health Ecosystem API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List user appointments
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update or cancel appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
  /ai-chat/sessions:
    get:
      operationId: listChatSessions
      tags: [ai-chat]
      summary: List AI chat sessions
      responses:
        "200":
          description: Chat sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
    post:
      operationId: createChatSession
      tags: [ai-chat]
      summary: Start new AI chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionInput"
      responses:
        "201":
          description: Chat session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /ai-chat/sessions/{id}/messages:
    get:
      operationId: getChatMessages
      tags: [ai-chat]
      summary: Get messages in a chat session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [ai-chat]
      summary: Send a message to the AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  # ─── HOSPITALS ────────────────────────────────────────────────────────────
...
  /blood-bank/donors:
    get:
      operationId: listBloodDonors
      tags: [blood-bank]
      summary: List blood donors
      parameters:
        - name: bloodGroup
          in: query
          schema:
            type: string
        - name: county
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Blood donors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BloodDonor"
    post:
      operationId: registerBloodDonor
      tags: [blood-bank]
      summary: Register as blood donor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodDonorInput"
      responses:
        "201":
          description: Donor registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodDonor"
  /blood-bank/requests:
    get:
      operationId: listBloodRequests
      tags: [blood-bank]
      summary: List blood requests
      responses:
        "200":
          description: Blood requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BloodRequest"
    post:
      operationId: createBloodRequest
      tags: [blood-bank]
      summary: Create blood request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodRequestInput"
      responses:
        "201":
          description: Blood request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"

  # ─── DASHBOARD (WOW ENDPOINTS) ────────────────────────────────────────────
...
components:
  schemas:
...
    Appointment:
      type: object
      required: [id, doctorId, doctorName, specialty, date, time, type, status]
      properties:
        id: { type: integer }
        doctorId: { type: integer }
        doctorName: { type: string }
        doctorImageUrl: { type: ["string", "null"] }
        specialty: { type: string }
        hospital: { type: string }
        date: { type: string }
        time: { type: string }
        type: { type: string }
        status: { type: string }
        notes: { type: ["string", "null"] }
        fee: { type: number }
    AppointmentInput:
      type: object
      required: [doctorId, date, time, type]
      properties:
        doctorId: { type: integer }
        date: { type: string }
        time: { type: string }
        type: { type: string }
        notes: { type: string }
    AppointmentUpdate:
      type: object
      properties:
        status: { type: string }
        date: { type: string }
        time: { type: string }
        notes: { type: string }
...
    ChatSession:
      type: object
      required: [id, topic, createdAt, messageCount]
      properties:
        id: { type: integer }
        topic: { type: string }
        createdAt: { type: string }
        messageCount: { type: integer }
        lastMessage: { type: ["string", "null"] }
    ChatSessionInput:
      type: object
      required: [topic]
      properties:
        topic: { type: string }
    ChatMessage:
      type: object
      required: [id, sessionId, role, content, createdAt]
      properties:
        id: { type: integer }
        sessionId: { type: integer }
        role: { type: string }
        content: { type: string }
        createdAt: { type: string }
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
...
    BloodDonor:
      type: object
      required: [id, name, bloodGroup, county, phone, lastDonation]
      properties:
        id: { type: integer }
        name: { type: string }
        bloodGroup: { type: string }
        county: { type: string }
        phone: { type: string }
        lastDonation: { type: ["string", "null"] }
        donationCount: { type: integer }
        available: { type: boolean }
    BloodDonorInput:
      type: object
      required: [bloodGroup, county, phone]
      properties:
        bloodGroup: { type: string }
        county: { type: string }
        phone: { type: string }
    BloodRequest:
      type: object
      required: [id, bloodGroup, unitsNeeded, hospital, urgency, status, createdAt]
      properties:
        id: { type: integer }
        bloodGroup: { type: string }
        unitsNeeded: { type: integer }
        hospital: { type: string }
        county: { type: string }
        urgency: { type: string }
        status: { type: string }
        patientName: { type: ["string", "null"] }
        contactPhone: { type: string }
        createdAt: { type: string }
    BloodRequestInput:
      type: object
      required: [bloodGroup, unitsNeeded, hospital, urgency, contactPhone]
      properties:
        bloodGroup: { type: string }
        unitsNeeded: { type: integer }
        hospital: { type: string }
        county: { type: string }
        urgency: { type: string }
        patientName: { type: string }
        contactPhone: { type: string }
...