# lib/api-zod/src/generated/types/bloodRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MAWIBO Super Health Ecosystem API
 * OpenAPI spec version: 0.1.0
 */

export interface BloodRequest {
  id: number;
  bloodGroup: string;
  unitsNeeded: number;
  hospital: string;
  county?: string;
  urgency: string;
  status: string;
  /** @nullable */
  patientName?: string | null;
  contactPhone?: string;
  createdAt: string;
}